changed float to double

+++ src/main/scala/com/sumologic/tools/costs/ec2_ri/optimizer/ec2/Ec2InstancesSummary.scala
@@ -3,7 +3,7 @@ package com.sumologic.tools.costs.ec2_ri.optimizer.ec2
 import org.json4s.DefaultFormats
 import org.json4s.native.Serialization.{read, write}
 
-case class Ec2InstancesSummary(familiesTotalSizes: Map[String, Float]) {
+case class Ec2InstancesSummary(familiesTotalSizes: Map[String, Double]) {
   // nothing
 
   def toJsonString: String = {
